chapter-09-content-pipeline/requirements.txt
crewai[tools]>=0.80.0
pydantic>=2.0.0
python-dotenv>=1.0.0
